basic form and template

+++ models/movie.go
@@ -6,7 +6,7 @@ import "gorm.io/gorm"
 type Movie struct {
     gorm.Model
     Title string
-    Rating int
+    Rating float32
 }
